toys: disable grep (egrep, fgrep) and iotop.
/system/bin is populated from 4 sources (at least: I don't claim
completeness).

Two of these are multi-call-
binaries:
  * toybox (which is this package)
  * toolbox (which is not completely build in one binary - i.e. grep
      is build from toolbox' sources, but is a standalone (aka external)
      command.

The other two possible sources are this two directories with source code
for standalone commands:
  * ${ANDROID_BUILD_TOP}/external
  * ${ANDROID_BUILD_TOP}/system/extras

While toolbox and toybox can be configured, the lineage/aosp build
system calls the module names for the standalone binaries implicitly.

Since there is no check if the symlinks of the multi-call-binaries
collide in any way with each other or the standalone commands, we (devs
and builders) have to take care of this.

This toybox modules are removed:

  * grep (incl. egrep and fgrep): use toolbox' standalone grep binary
  * iotop: use ${ANDROID_BUILD_TOP}/system/extras/iotop

Change-Id: I4ff1d6a0ec8c467d824618c2f343d7044d9acf4c